add auth to Course, refactor packages
