+++ src/main/java/cf/youngauthentic/consultant/model/department/DepartmentWithCourse.java
@@ -1,5 +1,6 @@
-package cf.youngauthentic.consultant.model;
+package cf.youngauthentic.consultant.model.department;
 
+import cf.youngauthentic.consultant.model.course.CourseWithoutTeachers;
 import org.springframework.data.rest.core.config.Projection;
 
 import java.util.List;
