+++ src/main/java/cf/youngauthentic/consultant/service/CourseService.java
@@ -1,8 +1,8 @@
 package cf.youngauthentic.consultant.service;
 
-import cf.youngauthentic.consultant.model.CourseEntity;
-import cf.youngauthentic.consultant.model.CourseEntityPK;
-import cf.youngauthentic.consultant.model.CourseWithoutTeachers;
+import cf.youngauthentic.consultant.model.course.CourseEntity;
+import cf.youngauthentic.consultant.model.course.CourseEntityPK;
+import cf.youngauthentic.consultant.model.course.CourseWithoutTeachers;
 import cf.youngauthentic.consultant.repo.CourseRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -35,7 +35,8 @@ public class CourseService {
         return courseRepo.findById(pk);
     }
 
-    public CourseEntity getCourseWithTeachers(int departmentId, int courseId) {
+    public CourseEntity getCourseWithTeachers(int departmentId, int courseId, String token) throws AuthException {
+        loginService.hasAuth(token, Auth.STUDENT);
         return courseRepo.findByDepartmentIdAndCourseId(departmentId, courseId);
     }
 }
